pyproject.toml
[tool.poetry]
name = "fuzy-jon"
version = "0.0.0"
description = ""
authors = ["lucemia <lucemia@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"

[tool.poetry.group.test.dependencies]
syrupy = "^4.0.2"
pytest-cov = "^4.1.0"
coveralls = "^3.3.1"
pytest-recording = "^0.12.2"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.5.0"
mypy = "^1.6.1"

[tool.poetry-dynamic-versioning]
enable = true
pattern = "default-unprefixed"

[build-system]
requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning"]
build-backend = "poetry_dynamic_versioning.backend"
